+++ test/pagy/extras/out_of_range_test.rb
@@ -29,10 +29,15 @@ describe Pagy do
 
   describe "#initialize" do
 
-    it 'initializes with out of range page' do
+    it 'works in :last_page mode' do
       pagy.must_be_instance_of Pagy
       pagy.page.must_equal pagy.last
       pagy.vars[:page].must_equal 100
+      pagy.offset.must_equal 100
+      pagy.items.must_equal 3
+      pagy.from.must_equal 101
+      pagy.to.must_equal 103
+      pagy.prev.must_equal 10
     end
 
     it 'raises OutOfRangeError in :exception mode' do
@@ -49,15 +54,21 @@ describe Pagy do
       pagy.prev.must_equal pagy.last
     end
 
+    it 'raises ArgumentError' do
+      proc { Pagy.new(vars.merge(out_of_range_mode: :unknown)) }.must_raise ArgumentError
+    end
+
   end
 
   describe "#series singleton for :empty_page mode" do
+
     it 'computes series for last page' do
       pagy = Pagy.new(vars.merge(out_of_range_mode: :empty_page))
       series = pagy.series
       series.must_equal [1, 2, 3, :gap, 9, 10, 11]
       pagy.page.must_equal 100
     end
+
   end
 
 end